Implement Toolbar and Menu Bar for OSX
